werror gantry
