+++ gantry/firmware/can_task.cpp
@@ -9,18 +9,23 @@
 #include "can/core/message_handlers/move_group_executor.hpp"
 #include "can/core/message_writer.hpp"
 #include "can/core/messages.hpp"
-#include "can/firmware/hal_can_bus.hpp"
-#include "can/firmware/hal_can_message_buffer.hpp"
 #include "common/core/freertos_message_queue.hpp"
 #include "common/core/freertos_task.hpp"
-#include "common/firmware/can.h"
 #include "common/firmware/errors.h"
 #include "common/firmware/spi_comms.hpp"
 #include "gantry/core/axis_type.hpp"
 #include "motor-control/core/linear_motion_system.hpp"
 #include "motor-control/core/motor.hpp"
 #include "motor-control/core/motor_messages.hpp"
+#pragma GCC diagnostic push
+// NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
+#pragma GCC diagnostic ignored "-Wvolatile"
+#include "can/firmware/hal_can_bus.hpp"
+#include "can/firmware/hal_can_message_buffer.hpp"
+#include "common/firmware/can.h"
+#include "motor_hardware.h"
 #include "platform_specific_hal_conf.h"
+#pragma GCC diagnostic pop
 
 using namespace freertos_task;
 using namespace freertos_can_dispatch;
@@ -45,91 +50,6 @@ static freertos_message_queue::FreeRTOSMessageQueue<Move> motor_queue(
 static freertos_message_queue::FreeRTOSMessageQueue<Ack> complete_queue(
     "Complete Queue");
 
-void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock enable */
-        __HAL_RCC_SPI2_CLK_ENABLE();
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        __HAL_RCC_GPIOB_CLK_ENABLE();
-        __HAL_RCC_GPIOC_CLK_ENABLE();
-        /**SPI2 GPIO Configuration
-        PB12  ------> SPI2_CS
-        PB13  ------> SPI2_SCK
-        PB14  ------> SPI2_MISO
-        PB15  ------> SPI2_MOSI
-        Step/Dir
-        PB1   ------> Motor Dir Pin
-        PC8   ------> Motor Step Pin
-        Enable
-        PA9   ------> Motor Enable Pin
-        */
-        GPIO_InitStruct.Pin = GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Chip select
-        GPIO_InitStruct.Pin = GPIO_PIN_12;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Dir
-        GPIO_InitStruct.Pin = GPIO_PIN_1;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Step
-        GPIO_InitStruct.Pin = GPIO_PIN_8;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOC,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-                      &GPIO_InitStruct);
-
-        // Enable
-        GPIO_InitStruct.Pin = GPIO_PIN_9;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOA,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-                      &GPIO_InitStruct);
-    }
-}
-SPI_HandleTypeDef hspi2 = {
-    .Instance = SPI2,
-    .Init = {.Mode = SPI_MODE_MASTER,
-             .Direction = SPI_DIRECTION_2LINES,
-             .DataSize = SPI_DATASIZE_8BIT,
-             .CLKPolarity = SPI_POLARITY_HIGH,
-             .CLKPhase = SPI_PHASE_2EDGE,
-             .NSS = SPI_NSS_SOFT,
-             .BaudRatePrescaler = SPI_BAUDRATEPRESCALER_32,
-             .FirstBit = SPI_FIRSTBIT_MSB,
-             .TIMode = SPI_TIMODE_DISABLE,
-             .CRCCalculation = SPI_CRCCALCULATION_DISABLE,
-             .CRCPolynomial = 7,
-             .CRCLength = SPI_CRC_LENGTH_DATASIZE,
-             .NSSPMode = SPI_NSS_PULSE_DISABLE}
-
-};
-
-/**
- * @brief SPI MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param hspi: SPI handle pointer
- * @retval None
- */
-void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock disable */
-        __HAL_RCC_SPI2_CLK_DISABLE();
-        HAL_GPIO_DeInit(GPIOB, GPIO_PIN_12 | GPIO_PIN_13 | GPIO_PIN_14 |
-                                   GPIO_PIN_15 | GPIO_PIN_1);
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_9);
-        HAL_GPIO_DeInit(GPIOC, GPIO_PIN_8);
-    }
-}
 spi::SPI_interface SPI_intf = {
 
     .SPI_handle = &hspi2,
@@ -210,8 +130,8 @@ static auto dispatcher = Dispatcher(
     if (MX_FDCAN1_Init(&fdcan1) != HAL_OK) {
         Error_Handler();
     }
-    __HAL_RCC_SPI2_CLK_ENABLE();
-    if (HAL_SPI_Init(&hspi2) != HAL_OK) {
+
+    if (initialize_spi() != HAL_OK) {
         Error_Handler();
     }
     can_bus::setup_node_id_filter(can_bus_1, axis_type::get_node_id());
